requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
django-heroku==0.3.1
django-taggit==2.1.0
gunicorn==20.1.0
Pillow==9.0.1
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.2.0
